% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_export_workbook.R
\name{create_export_workbook}
\alias{create_export_workbook}
\title{Create export workbook}
\usage{
create_export_workbook(
  site_names,
  feature_names,
  action_names,
  site_data,
  site_status_data,
  feature_data,
  action_expectation_data,
  summary_results_data,
  site_results_data,
  feature_results_data,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{feature_names}{\code{character} names of biodiversity features. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{action_names}{\code{character} names of management actions. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{site_data}{\code{data.frame} containing site data.}

\item{site_status_data}{\code{data.frame}containing site status data.}

\item{feature_data}{\code{data.frame} containing feature data.}

\item{action_expectation_data}{\code{data.frame} containing expectation data.}

\item{summary_results_data}{\code{data.frame} containing summary results data.}

\item{site_results_data}{\code{data.frame} containing site results data.}

\item{feature_results_data}{\code{data.frame} containing feature results data.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{Workbook} object.
}
\description{
This function creates an Excel Workbook with results.
}
