% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_workbook.R
\name{create_project_workbook}
\alias{create_project_workbook}
\title{Create data workbook}
\usage{
create_project_workbook(
  site_ids,
  site_descriptions,
  feature_ids,
  feature_descriptions,
  action_ids,
  action_descriptions,
  site_data,
  feasibility_data,
  feature_data,
  action_expectation_data,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} identifiers for sites.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{site_descriptions}{\code{character} descriptions of sites.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{feature_ids}{\code{character} identifiers for biodiversity features.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{feature_descriptions}{\code{character} descriptions of biodiversity
features.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{action_ids}{\code{character} identifiers for management actions.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{action_descriptions}{\code{character} descriptions of management actions.
No missing (\code{NA}) values are permitted. This object must contain at least
one value.}

\item{site_data}{\code{data.frame} containing site data.}

\item{feasibility_data}{\code{data.frame}containing feasibility data.}

\item{feature_data}{\code{data.frame} containing feature data.}

\item{action_expectation_data}{\code{list} of \code{data.frame} objects
containing expectation data.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{Workbook} object.
}
\description{
This function creates an Excel Workbook with data.
}
