% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template_workbook.R
\name{create_template_workbook}
\alias{create_template_workbook}
\title{Create template workbook}
\usage{
create_template_workbook(site_names, feature_names, action_names, parameters)
}
\arguments{
\item{site_names}{\code{character} names of sites. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{feature_names}{\code{character} names of biodiversity features. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{action_names}{\code{character} names of management actions. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{Workbook} object.
}
\description{
This function creates a data template Excel Workbook.
}
