% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_error_data}
\alias{format_error_data}
\title{Format error data}
\usage{
format_error_data(site_ids, feature_ids, action_ids, problem, parameters)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{problem}{\code{ConservationProblem} object.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{list} object contain \code{data.frame} objects summarizing the results.
}
\description{
Format results for a prioritization problem that threw an error.
}
