% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_locked_data}
\alias{format_locked_data}
\title{Format locked data}
\usage{
format_locked_data(
  site_ids,
  feature_ids,
  action_ids,
  site_status_data,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{site_status_data}{\code{data.frame} with site status data.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with locked data for prioritizations.
}
\description{
Format locked data
}
