% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_pu_data}
\alias{format_pu_data}
\title{Format planning unit data}
\usage{
format_pu_data(
  site_ids,
  feature_ids,
  action_ids,
  site_data,
  action_expectation_data,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{site_data}{\code{data.frame} with site data.}

\item{action_expectation_data}{\code{list} of \code{data.frame} objects with
the expected amount of each feature given each action.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with planning unit data for prioritizations.
}
\description{
Format planning unit data
}
