% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_target_data}
\alias{format_target_data}
\title{Format target data}
\usage{
format_target_data(site_ids, feature_ids, action_ids, feature_data, parameters)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{feature_data}{\code{data.frame} with feature data.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with target data for prioritizations.
}
\description{
Format target data
}
