% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_weights_data}
\alias{format_weights_data}
\title{Format weights data}
\usage{
format_weights_data(
  site_ids,
  feature_ids,
  action_ids,
  feature_data,
  parameters
)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{feature_data}{\code{data.frame} with feature data.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with weights data for prioritizations.
}
\description{
Format weights data
}
