% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_zone_data}
\alias{format_zone_data}
\title{Format zone data}
\usage{
format_zone_data(site_ids, feature_ids, action_ids, parameters)
}
\arguments{
\item{site_ids}{\code{character} identifiers of sites.}

\item{feature_ids}{\code{character} identifiers of features.}

\item{action_ids}{\code{character} identifiers of actions.}

\item{parameters}{\code{list} object with configuration parameters.}
}
\value{
\code{data.frame} with zone data for prioritizations.
}
\description{
Format zone data
}
