% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_spreadsheet.R
\name{is_valid_spreadsheet}
\alias{is_valid_spreadsheet}
\title{Is validate spreadsheet?}
\usage{
is_valid_spreadsheet(x, parameters)
}
\arguments{
\item{x}{\code{character} file path.}

\item{parameters}{\code{list} configuration parameters.}
}
\value{
\code{logical} indicating if file path contains valid data.
}
\description{
Verify if an Excel Spreadsheet is in the correct format.
}
