% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_configuration.R
\name{read_data_configuration}
\alias{read_data_configuration}
\title{Read data configuration}
\usage{
read_data_configuration()
}
\value{
\code{list} object with parameters.
}
\description{
Load data format configuration parameters.
}
