% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spreadsheet_data.R
\name{read_spreadsheet_data}
\alias{read_spreadsheet_data}
\title{Read spreadsheet data}
\usage{
read_spreadsheet_data(x, parameters)
}
\arguments{
\item{x}{\code{character} file path.}

\item{parameters}{\code{list} configuration parameters for}
}
\value{
\code{list} object.
}
\description{
Read prioritization input data from an Excel Spreadsheet. Note that
this function assumes that the data are formatted correctly, following
the NCC Data App.
}
