% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_project_data.R
\name{simulate_project_data}
\alias{simulate_project_data}
\title{Simulate project data}
\usage{
simulate_project_data(
  n_sites,
  n_features,
  n_actions,
  parameters,
  prop_locked_out = 0.05
)
}
\arguments{
\item{n_sites}{\code{integer} number of sites.}

\item{n_features}{\code{integer} number of features.}

\item{n_actions}{\code{integer} number of actions.}

\item{parameters}{\code{list} object containing configuration parameters.}

\item{prop_locked_out}{\code{numeric} proportion of locked out actions. Defaults
to 0.05 (5\%).}
}
\value{
\code{list} object containing the simulated data.
}
\description{
Simulate project data
}
