% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_data.R
\name{template_action_expectation_data}
\alias{template_action_expectation_data}
\title{Template action expectation data}
\usage{
template_action_expectation_data(
  site_names,
  feature_names,
  action_name,
  parameters
)
}
\arguments{
\item{site_names}{\code{character} names of sites. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{feature_names}{\code{character} names of biodiversity features. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{action_name}{\code{character} name of the action.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{data.frame} object.
}
\description{
Create action expectation data for the template workbook.
}
