% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_data.R
\name{template_feature_data}
\alias{template_feature_data}
\title{Template feature data}
\usage{
template_feature_data(feature_names, parameters)
}
\arguments{
\item{feature_names}{\code{character} names of biodiversity features. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{data.frame} object.
}
\description{
Create feature data for the template workbook.
}
