% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_data.R
\name{template_site_data}
\alias{template_site_data}
\title{Template site data}
\usage{
template_site_data(site_names, action_names, parameters)
}
\arguments{
\item{site_names}{\code{character} names of sites. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{action_names}{\code{character} names of management actions. No missing
(\code{NA}) values are permitted. This object must contain at least one
name.}

\item{parameters}{\code{list} object containing parameters to customize
appearance of worksheet.}
}
\value{
\code{data.frame} object.
}
\description{
Create site data for the template workbook.
}
