% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{whatdata}
\alias{whatdata}
\title{whatdata: Import and export management action data}
\description{
The package provides functions for importing and exporting management action
data. It provides consistent data format standards for the
\href{https://github/NCC-CNC/whattodo}{What To Do} and
\href{https://github/NCC-CNC/whattemplatemaker}{What Template Maker} web
applications. It is not designed for general purpose use outside of these
applications.
}
